requirements.txt
annoy>=1.8.3
Pillow>=9.0.1
matplotlib>=2.0.2
numpy>=1.14.5
pandas>=0.22.0
pygame>=1.9.3
PyOpenGL>=3.1.0
scipy>=0.19.0
scikit-image>=0.13.0
bokeh==1.0.4
kubernetes>=8.0.0b1,<=8.0.1
redis>=2.10.6
minio>=4.0.5
pytest>=3.8.2
psutil>=5.5.0
joblib>=0.17.0
protobuf==3.20.*
ipython
h5py==2.10.0
opencv-python==4.11.0.86
